+++ android/src/main/kotlin/com/jayfinava/flutteredgedetection/EdgeDetectionHandler.kt
@@ -37,6 +37,8 @@ class EdgeDetectionHandler : MethodCallHandler, PluginRegistry.ActivityResultLis
         const val AUTO_CAPTURE_PREVIEW_BUTTON_VERTICAL_PADDING = "auto_capture_preview_button_vertical_padding"
         const val AUTO_CAPTURE_PREVIEW_BUTTON_BACKGROUND_COLOR = "auto_capture_preview_button_background_color"
         const val AUTO_CAPTURE_PREVIEW_BUTTON_BORDER_RADIUS = "auto_capture_preview_button_border_radius"
+        const val AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_TEXT =
+            "auto_capture_preview_retake_button_text"
         const val AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_TEXT_COLOR =
             "auto_capture_preview_retake_button_text_color"
         const val AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_TEXT_SIZE =
@@ -49,6 +51,8 @@ class EdgeDetectionHandler : MethodCallHandler, PluginRegistry.ActivityResultLis
             "auto_capture_preview_retake_button_background_color"
         const val AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_BORDER_RADIUS =
             "auto_capture_preview_retake_button_border_radius"
+        const val AUTO_CAPTURE_PREVIEW_NEXT_BUTTON_TEXT =
+            "auto_capture_preview_next_button_text"
         const val AUTO_CAPTURE_PREVIEW_NEXT_BUTTON_TEXT_COLOR =
             "auto_capture_preview_next_button_text_color"
         const val AUTO_CAPTURE_PREVIEW_NEXT_BUTTON_TEXT_SIZE =
@@ -225,6 +229,10 @@ class EdgeDetectionHandler : MethodCallHandler, PluginRegistry.ActivityResultLis
                 previewButtonBorderRadius
             )
             putString(
+                AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_TEXT,
+                call.argument<String>(AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_TEXT)
+            )
+            putString(
                 AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_TEXT_COLOR,
                 call.argument<String>(AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_TEXT_COLOR)
                     ?: previewButtonTextColor
@@ -253,6 +261,10 @@ class EdgeDetectionHandler : MethodCallHandler, PluginRegistry.ActivityResultLis
                 AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_BORDER_RADIUS,
                 call.argument<Number>(AUTO_CAPTURE_PREVIEW_RETAKE_BUTTON_BORDER_RADIUS)?.toDouble()
                     ?: previewButtonBorderRadius
+            )
+            putString(
+                AUTO_CAPTURE_PREVIEW_NEXT_BUTTON_TEXT,
+                call.argument<String>(AUTO_CAPTURE_PREVIEW_NEXT_BUTTON_TEXT)
             )
             putString(
                 AUTO_CAPTURE_PREVIEW_NEXT_BUTTON_TEXT_COLOR,